refactor: patient page and circular indicator
